Finish all service and controller unit tests

+++ Project.Api/Controllers/EmployeesController.cs
@@ -1,11 +1,12 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Project.Application.Dtos.Employee;
 using Project.Core.Entities;
 using Project.Services;
 using System;
-using System.Collections.Generic;
 using System.Linq;
+using System.Linq.Expressions;
 using System.Threading.Tasks;
 
 namespace Project.Api.Controllers
@@ -24,38 +25,100 @@ namespace Project.Api.Controllers
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetAsync()
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        public async Task<IActionResult> GetAsync([FromQuery] int pageIndex = 1, int pageSize = 10, string searchValue = null, string orderBy = null, bool isDelete = false)
         {
             try
             {
-                IEnumerable<Employee> employees = await _employeeService.GetAllAsync();
+                Expression<Func<Employee, bool>> filter = null;
+                Func<IQueryable<Employee>, IOrderedQueryable<Employee>> order = null;
+                string include = string.Empty;
+
+                if (!string.IsNullOrWhiteSpace(searchValue))
+                {
+                    filter = e => e.Id.ToString().Equals(searchValue)
+                            || e.FirstName.Contains(searchValue)
+                            || e.LastName.Contains(searchValue)
+                            || e.Phone.Contains(searchValue)
+                            || e.Email.Contains(searchValue);
+                }
+                if (!string.IsNullOrWhiteSpace(orderBy))
+                {
+                    switch (orderBy)
+                    {
+                        case "id_desc":
+                            order = x => x.OrderByDescending(e => e.Id);
+                            break;
+                        case "firstName":
+                            order = x => x.OrderBy(e => e.FirstName);
+                            break;
+                        case "firstName_desc":
+                            order = x => x.OrderByDescending(e => e.FirstName);
+                            break;
+                        case "lastName":
+                            order = x => x.OrderBy(e => e.LastName);
+                            break;
+                        case "lastName_desc":
+                            order = x => x.OrderByDescending(e => e.LastName);
+                            break;
+                        case "phone":
+                            order = x => x.OrderBy(e => e.Phone);
+                            break;
+                        case "phone_desc":
+                            order = x => x.OrderByDescending(e => e.Phone);
+                            break;
+                        case "dob":
+                            order = x => x.OrderBy(e => e.DateOfBirth);
+                            break;
+                        case "dob_desc":
+                            order = x => x.OrderByDescending(e => e.DateOfBirth);
+                            break;
+                        default:
+                            order = x => x.OrderBy(e => e.Id);
+                            break;
+                    }
+                }
+
+                var employees = await _employeeService.GetAsync(pageIndex, pageSize, filter: filter, orderBy: order, include, isDelete);
                 return Ok(employees.Select(e => _mapper.Map<EmployeeDto>(e)));
             }
             catch (Exception ex)
             {
-                return StatusCode(500, ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
             }
         }
 
         [HttpGet("{id}", Name = "GetEmployeeById")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<IActionResult> GetByIdAsync(string id)
         {
             try
             {
-                Employee employee = await _employeeService.GetByIdAsync(id);
-                if (employee is null)
+                if (string.IsNullOrWhiteSpace(id)
+                    || !Guid.TryParse(id, out _))
                 {
-                    return NotFound();
+                    return BadRequest();
                 }
-                return Ok(_mapper.Map<EmployeeDto>(employee));
+
+                var employee = await _employeeService.GetByIdAsync(id);
+                return employee is null
+                    ? NotFound()
+                    : Ok(_mapper.Map<EmployeeDto>(employee));
             }
             catch (Exception ex)
             {
-                return StatusCode(500, ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
             }
         }
 
         [HttpPost]
+        [ProducesResponseType(StatusCodes.Status201Created)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<IActionResult> AddAsync([FromBody] CreateEmployeeDto dto)
         {
             try
@@ -65,33 +128,35 @@ namespace Project.Api.Controllers
                     return BadRequest();
                 }
 
-                Employee entity = _mapper.Map<Employee>(dto);
+                var entity = _mapper.Map<Employee>(dto);
 
                 await _employeeService.AddAsync(entity);
                 return CreatedAtAction(nameof(GetByIdAsync), new { id = entity.Id }, dto);
             }
             catch (Exception ex)
             {
-                return StatusCode(500, ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
             }
         }
 
         [HttpPut("{id}")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<IActionResult> UpdateAsync(string id, [FromBody] UpdateEmployeeDto dto)
         {
             try
             {
-                if (!ModelState.IsValid)
+                if (string.IsNullOrWhiteSpace(id)
+                    || !Guid.TryParse(id, out _)
+                    || !ModelState.IsValid
+                    || !(id == (dto.Id)))
                 {
                     return BadRequest();
                 }
 
-                if (!id.Equals(dto.Id))
-                {
-                    return BadRequest();
-                }
-
-                Employee inDatabase = await _employeeService.GetByIdAsync(id);
+                var inDatabase = await _employeeService.GetByIdAsync(id);
                 if (inDatabase is null)
                 {
                     return NotFound();
@@ -103,35 +168,61 @@ namespace Project.Api.Controllers
             }
             catch (Exception ex)
             {
-                return StatusCode(500, ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
             }
         }
 
         [HttpDelete("{id}")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<IActionResult> DeleteAsync(string id)
         {
             try
             {
+                if (string.IsNullOrWhiteSpace(id)
+                    || !Guid.TryParse(id, out _))
+                {
+                    return BadRequest();
+                }
+
                 await _employeeService.DeleteAsync(id);
                 return NoContent();
             }
+            catch (ArgumentNullException)
+            {
+                return NotFound();
+            }
             catch (Exception ex)
             {
-                return StatusCode(500, ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
             }
         }
 
         [HttpDelete("{id}/Delete")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<IActionResult> HardDeleteAsync(string id)
         {
             try
             {
+                if (string.IsNullOrWhiteSpace(id)
+                    || !Guid.TryParse(id, out _))
+                {
+                    return BadRequest();
+                }
+
                 await _employeeService.HardDeleteAsync(id);
                 return NoContent();
             }
+            catch (ArgumentNullException ex)
+            {
+                return StatusCode(StatusCodes.Status404NotFound, ex.Message);
+            }
             catch (Exception ex)
             {
-                return StatusCode(500, ex.Message);
+                return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
             }
         }
     }
